// src/types/user.ts
/** ユーザー */
export type User = {
  /** id */
  id: string;
  /** 名前 */
  name: string;
  /** 郵便番号 */
  postalCode: string;
  /** 住所 */
  address: string;
  /** 電話番号 */
  phoneNumber: string;
  /** メールアドレス */
  email: string;
  /** パスワード */
  password: string;
};